Validate numeric command line arguments
The numeric command line arguments are currently not validated. If a
user is passing a string by accident as an argument, or is passing an
invalid argument. This will not be detected by the application. Checking
that the arguments make sense when they are passed would informative to
a user. For example, when ``--buttons-per-row 0`` or if
``--buttons-per-row o`` is passed as an argument, wlogout will crash due
to division by 0 later during execution. This patch will add validation
to the numeric arguments.

This validation comes in 2 phases:

- We check that the argument passed by the user is a number.
  This is currently done with ``atoi()``, but this function returns 0 if
  the string is not a number. Instead, swap it for ``sscanf()`` and try to
  parse and ``%d``. This will fail if a a string which is not a number is
  passed as an argument. The new ``parse_numarg()`` function will take
  care of this.
- The parsed values are checked to ensure they are within the expected
  ranges before resuming. Mostly this means that we check that values
  are not negative as this is what the documentation suggests. And in
  the case of ``buttons_per_row``, we need a value of at least 1.

We try to provide informative feedback to the user when an argument
fails validation. To make this somewhat simpler, we also add an array of
names for spacing and margins to make writing their names over and over
less cumbersome. And whilst we're here, also update the documentation for
the ``--buttons-per-row`` option to be at least 1.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>