Add named constants for the margins
This makes it easier to understand which entry of the ``margin`` array
corresponds to a given direction. We also update all uses of the
``margin`` array to use the named constants for indexing. Whilst we are
here, we also make the change to only do the string to int conversion once
for the ``--margin option``. There is no functional changes to the
application apart of this patch.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>