Statically allocate the buttons array
There is currently no means by a user to change the number of buttons.
This is fixed to an upper limit of 100 in the ``default_size`` variable.
Though as the ``default_size`` is not a constant expression, it cannot
be used to define a fixed-size array in the global scope. Since the
``default_size`` never changes, we can substitute it for a constant
(macro) which enables us to define the ``buttons`` as a fixed size array
and eliminating the need to allocate memory dynamically. This makes
sense as this is what the application will be doing anyways. There is no
functional change intended by this patch.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>