Abort if dynamic memory allocation fails
There are quite a few places where dynamic memory allocation is used
throughout wlogout. Not every allocation is checked for success.
Although it is very unlikely that a ``malloc()`` call should fail, if it
does then program should cease execution and fail fast. So this patch
adds the ``abort_if()`` function which will exit with a failure if we
are running in an abnormal state. We let the system take care of the
resource cleanup in this case.

Signed-off-by: John Berg <john.berg.96bits@gmail.com>